Change of bitshift in MIDI function to comply with new compiler behaviour
Bitshift is different for INT und UNIT. It has been found out that
eihter the newer compiler treated the bitshift based on the source type
(INT) rather than the target type or that the different behaviour of
bitshift on UNIT and INT was introduced somewhere between avr-gcc
versions 4.8.2 and 11.1. No forensics has been done on this but the
behaviour in both cases has been reproduced and the new changed code
enforces logic bitshift (e.g. UNIT behaviour).