New features soft bend, detune tarcking; unison fixing
The setting "spread" is not a patch parameter (777). The load of the
default patch is now on the Preset button and the local on/off has moved
from "7" to "0".

Various fixes in the unison pedal latch and related assigner logic
(missing key off event). Also: on pedal press the played chord is now
retriggered. The pedal press is not sent during unison mode. An the
incoming MIDI switch to unison is only carried out if not in unison
mode.

A new parameter detune tracking (888).

In order to test softer mod wheels the mod wheel range contains now 8
setting (4 soft ones and 4 original ones).