Introduction of "pick-me-up" pot behaviour in preset panel mode
As described in isse #26: the remedy for the situation in which the pots
(and switches) don't show the current active panel parameters is a
pick-me-up mode in which the user needs to turn a pot until is
corresponds to the active value before it takes effect (e.g. the uer
picks up the value). In order to give feedback to the user the disply
show whether the pot value is below or above the current active control
value and then switches to numeric display once picked up. The pot
status is reset if a patch is loaded or a MIDI patch SysEs is loaded
into active controls because these are the situations where the panel
doesn't show the active values.

The implementation here is only partial:
- Only supported in preset panel mode (e.g. with Preset and TapeTo
activated)
- Only pots are supported, not switches

Note: the following pots are not part of this logic, because they are
not stored in a patch: Mod Wheel, Bender, Main Tune, Main Volume, Speed