Potmux change detect mod and UI logic correction
In Potmux the pot value is not forced to the maximum, when a certain
threshold is reached. It is unlcear whether that is a sound solution, so
it's a test for solving the issue #52 when the filter sustain maximum is
notably smaller than the peak of the attack/decay curve.

In addition to that there are the following changes:
- resetting of MIDI pitch bend (included in bender calibration)
- Fixes for the UI LED logic
- Small changes (nice display messages, no prog change MIDI when storing
a patch)