Fix for analog clock in (#109) and new reduced mod wheel (#110)
Tape in (clock) is now read in the 2kHz cycle and not in the 500Hz
cycle.

Note: somehow the error that was observed is a hint that not all cycles
are properly completed (already in v2.1RC3). This requires further
examination. But the fix for clock works anyway.

The mod wheel ist softer no with a lower maximum value. For target
vibration->VCA ist is the same as before (VCA is a much weaker
modulation effect).