Fix for mod delay glitches (#180)
The parameter value modulation delay is now cut off below about 20, i.e.
not delay below that and jump to the delay value corresponding to 20.
The reason is that the corresponding time frame below 20 is too short
for the update cycle of the Prophet 600. That produced glitches. See
also mod delay reset issue #145.