Wrong vib for <v2022 patches (#192)
There was a factor (division) of 4 missing: this was implemented in the
v2.* (double bit shift >>2) but no longer in the v2022 version. In the
patch conversion the factor was missing. Also, there was a small bug in
relation to the DEADBAND, which affect the LFO and VIB amount conversion
(although this was probably not noticable). It is corrected now.